lab01/app/requirements.txt
Flask==3.1.0
PyJWT==2.10.1
python-dotenv==1.0.0
